adding a step in build_buildroot.sh to prepare the kernel headers for external module compilation (#10059)
adding file list to push into minio for linux-headers

1. merging mnic with latest ionic
2. added a script to prepare kernel headers when user run
build_buildroot.sh
3. making Linux driver Makefile compatible to compile for ARCM and for
X86

added files which forgot to add in previous commit

pushed the linux-headers for mnic compilation

fixing review comments

using make variable to specify NIC directory

fixed some of the review comments and checkpatch.pl

fixing some more review comments

one more small fix

fixed some more review comments

reverting the buildroot minio tag as suggested by stavros

+++ nic/tools/sysinit.sh
@@ -57,6 +57,23 @@ if [[ ! -f $PLATFORM_DIR/bin/nicmgrd ]]; then
     exit 1
 fi
 
+if [[ ! -f $PLATFORM_DIR/drivers/ionic_mnic.ko ]]; then
+    echo "mnic driver not found"
+    exit 1
+fi
+
+if [[ ! -f $PLATFORM_DIR/drivers/mnet.ko ]]; then
+    echo "mnet driver not found"
+    exit 1
+fi
+
+#remove the old drivers if they exist
+#TODO: Need to remove this rmmod once we remove mnic from plkatform-linux tree
+rmmod mnet ionic_mnic
+
+insmod $PLATFORM_DIR/drivers/ionic_mnic.ko
+insmod $PLATFORM_DIR/drivers/mnet.ko
+
 # start sysmgr
 rm -f *.log
 rm -f agent.log* /tmp/*.db
